openhurricane/requirements.txt
python-keystoneclient
python-novaclient
python-glanceclient
pytest
Faker
pyyaml